Changes for build 3.0.3
1. PR-32135  Retest HUC-4 export from aggregated HUC8 where HUC4s have an ending date.

Had to fix SQL for fetching unique list of hucs.
And had to remove a duplicate huc4 from reference list huc.
IN, IL, WI, and MI would be impacted.

2. Exit AWUDS if access is none.

Change made:

If not in GS domain then auto exit, see no GUI.

If in GS domain, see GUI, have to prompt for state first. And if the user still does not have access
for any reason then exit GUI.

3. PR32139 Add link to Google form for citation requests on the assign citations form.

 From Nancy:  OK, I've changed it to "anyone with the link can edit", but they do have to be logged in with a DOI login.

4. PR-1b
State Radio button on AREA tab

Change made, and email sent to Kathie to request approval from WUUG to disable this option:

I don't see much point in doing a QA Report for State-level datasets, given the amount of use we expect.  In 
order to do the QA Report for State datasets someone would need to go through the data dictionary and provide 
the programmer a list of equations that would be used in the report. The report would need programmed and then 
tested. For now, to avoid any confusion, the state selection under the quality-assurance program is being made 
invisible for release 3.0.
                    '
                    rdoAreaUnitState.Visible = False

5. PR-32090 Interactive Data Entry - Weird rounding behavior when rounding to 2 decimal places  

Fixed, had wrong value in a call to the rounding routine.

6. PR32052

State trend report now outputs state totals for 1960-1980, and does state sums from 1985-2015.

Also, we will not do multi-state trends for 3.0 only single state trends.
Maybe we need a new PR to request multi-state trends and then close PR 32052.

7. PR32131

Temporal NOW is now translated as 9999. So areas beginning in 2016 should now be allowed in 2016 datasets. 
Because county 102 now has the date range 2016 to 9999, instead of 2016 to 2015.

I deleted all the 2016 datasets because there internal data was bad. 
To see the changes you have to start over by creating new 2016 test datasets.

SD and Alaska are effected.