1. PR 32130 Changed the instructions for the log report to refer to selecting a dataset instead of selecting areas.
Checked other instructions and corrected them to be consistent with what is selected.

Made the Save and Retrieve buttons invisible if the do not apply.

Made the State area selection in all operation invisible if it does not apply.

2. PR 32125

Rewrote text as:

  lblInstructs.Text = "Choose an Area Type for the data storage option, select a data dictionary, enter year of 

data value and provide a dataset name. Click 'Execute' when done."
odd

3. PR 32120

Delete functionality was added to the citation form.