correct spelling and draft readme, fixed:
1. PR 28689

Reworked how nulls are displayed in BTBC. Blanks should now be displayed in the basic tables by 

category report if a calculation contains a null value.

2. PR 32075

When you click on the Generate Log Report button on the Area tab, you now get a dialog box that 

gives a default file name for the report and asks you where you want to output the report.