MSD failed on countyaquifer because a view had a bad data element name. View is fixed.
Also, the out of bound index error was due to a hard coded
array dim of 260 areas. Texas as >2700 countyaquifer  areas. 
Code was added to dynamically get the count of areas that exist for each State,
then the count is used to dimension the output array.
You may see a slight performance improvement for States
having <50 areas.