Datasets of 2 different years: NWULT feels that it's too likely to be confusing to have AWUDS do both "forward" and "backward" calculations depending on the order the user clicks on datasets of 2 different years--users often don't pay that much attention and then won't know why their results are inconsistent from one report to another.  Always use the earlier year as the "base", and use the following equation: Percent Change = ((Comparison - Base)/Base)*100
Place the base dataset's values in the column to the left (this had gotten inadvertently switched in the test version of AWUDS 3.0).

Datasets of the same year:
AWUDS should use the first dataset the user selects as the base, and the second as the comparison. Again put the base dataset values in the left column.